agent-canvas-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
greenlet>=3.0.3
aiosqlite>=0.20.0
pydantic>=2.7.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0
fastmcp>=2.0.0
kiwipiepy>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.1
sse-starlette>=2.1.0
aioredis>=2.0.1
